Fix the DiningDetailView
