+++ berkeley-mobile/Home/Dining/DiningDataSource/BMDiningLocation.swift
@@ -125,11 +125,32 @@ struct BMMealCategory: Codable, Hashable {
     let menuItems: [BMMenuItem]
 }
 
-struct BMMenuItemDetail: Codable, Hashable {
+struct BMMenuRecipeDetail: Codable, Hashable {
     let servingSize: String?
     let nutrition: [String: String]?
     let ingredients: String?
     let allergens: [String]?
+
+    var calories: Double {
+        Double(nutrition?["Calories (kcal)"] ?? "0") ?? 0
+    }
+
+    var remainingNutrition: [String: String] {
+        nutrition?.filter { $0.key != "Calories (kcal)" } ?? [:]
+    }
+
+    var protein: Int {
+        Int(nutrition?["Protein (g)"]?.filter { $0.isNumber } ?? "0") ?? 0
+    }
+
+    var carb: Int {
+        Int(nutrition?["Carbohydrate (g)"]?.filter { $0.isNumber } ?? "0") ?? 0
+    }
+
+    var fat: Int {
+        (Int(nutrition?["Total Lipid/Fat (g)"]?.filter { $0.isNumber } ?? "0") ?? 0) +
+        (Int(nutrition?["Trans Fat (g)"]?.filter { $0.isNumber } ?? "0") ?? 0)
+    }
 }
 
 struct BMMenuItem: Codable, Hashable {
@@ -138,7 +159,7 @@ struct BMMenuItem: Codable, Hashable {
     let menuId: String?
     let itemId: String?
     let dataLocation: String?
-    let recipeDetails: BMMenuItemDetail?
+    let recipeDetails: BMMenuRecipeDetail?
 }
 
 struct BMMealIcon: Codable, Hashable {
