+++ berkeley-mobile/Home/Dining/DiningDetailView.swift
@@ -197,148 +197,3 @@ struct DiningMenuItemIconsView: View {
         }
     }
 }
-
-// MARK: - DiningMenuItemDetailView
-
-struct DiningMenuItemDetailView: View {
-
-    var menuItem: BMMenuItem
-
-    var body: some View {
-        ScrollView {
-            VStack(alignment: .leading, spacing: 20) {
-                if let menuDetail = menuItem.recipeDetails {
-                    if let nutrition = menuDetail.nutrition, !nutrition.isEmpty {
-                        let calories = Double(nutrition["Calories (kcal)"] ?? "0") ?? 0
-                        let remainingNutrition = nutrition.filter { $0.key != "Calories (kcal)" }
-                        let protein = Int(nutrition["Protein (g)"]?.filter { $0.isNumber } ?? "0") ?? 0
-                        let carb = Int(nutrition["Carbohydrate (g)"]?.filter { $0.isNumber } ?? "0") ?? 0
-                        let fat = (Int(nutrition["Total Lipid/Fat (g)"]?.filter { $0.isNumber } ?? "0") ?? 0) + (Int(nutrition["Trans Fat (g)"]?.filter { $0.isNumber } ?? "0") ?? 0)
-                        
-                        if let servingSize = menuDetail.servingSize {
-                            VStack(alignment: .leading, spacing: 8) {
-                                Text("Serving Size")
-                                    .font(Font(BMFont.bold(20)))
-                                DiningDetailRowView {
-                                    Text(servingSize)
-                                        .font(Font(BMFont.regular(15)))
-                                }
-                            }
-                        }
-                        
-                        if !remainingNutrition.isEmpty {
-                            let macros: [(label: String, color: Color)] = [
-                                ("Protein", .green),
-                                ("Carb", .blue),
-                                ("Fat", .orange)
-                            ]
-                            VStack(alignment: .leading, spacing: 8) {
-                                Text("MacroNutrients")
-                                    .font(Font(BMFont.bold(20)))
-                                VStack(alignment: .leading, spacing: 8) {
-                                    Text("Calories: \(String(format: "%.1f", calories))")
-                                        .font(Font(BMFont.regular(15)))
-                                    ProgressCapsule(protein: protein, fat: fat, carb: carb)
-                                    HStack {
-                                        Spacer()
-                                        ForEach(macros, id: \.label) { macro in
-                                            HStack(spacing: 6) {
-                                                Circle()
-                                                    .fill(macro.color)
-                                                    .frame(width: 12, height: 12)
-                                                Text(macro.label)
-                                            }
-                                            Spacer()
-                                        }
-                                    }
-                                }
-                                .padding()
-                                .frame(maxWidth: .infinity)
-                                .shadowfy()
-                            }
-                        }
-
-
-                        if !remainingNutrition.isEmpty {
-                            VStack(alignment: .leading, spacing: 8) {
-                                Text("Nutrition Facts")
-                                    .font(Font(BMFont.bold(20)))
-                                VStack(spacing: 8) {
-                                    ForEach(remainingNutrition.sorted(by: { $0.key < $1.key }), id: \.key) { key, value in
-                                        DiningDetailRowView {
-                                            HStack {
-                                                Text(key)
-                                                    .font(Font(BMFont.regular(15)))
-                                                Spacer()
-                                                Text(value)
-                                                    .font(Font(BMFont.medium(15)))
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    if let ingredients = menuDetail.ingredients {
-                        VStack(alignment: .leading, spacing: 8) {
-                            Text("Ingredients")
-                                .font(Font(BMFont.bold(20)))
-                            Text(ingredients)
-                                .font(Font(BMFont.regular(15)))
-                        }
-                        .padding()
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                        .shadowfy()
-                    }
-
-                    if let allergens = menuDetail.allergens, !allergens.isEmpty {
-                        VStack(alignment: .leading, spacing: 8) {
-                            Text("Allergens")
-                                .font(Font(BMFont.bold(20)))
-                            DiningDetailRowView {
-                                Text(allergens.joined(separator: ", "))
-                                    .font(Font(BMFont.regular(15)))
-                            }
-                        }
-                    }
-                } else {
-                    Text("No Recipe Details")
-                        .font(Font(BMFont.regular(15)))
-                        .foregroundStyle(.secondary)
-                }
-            }
-            .padding()
-        }
-        .navigationTitle(menuItem.name)
-        .navigationBarTitleDisplayMode(.inline)
-    }
-}
-
-struct ProgressCapsule: View {
-    var protein: Int
-    var fat: Int
-    var carb: Int
-
-    var body: some View {
-        let total = protein + fat + carb
-        if total > 0 {
-            let segments: [(value: Int, color: Color)] = [
-                (protein, .green),
-                (carb, .blue),
-                (fat, .orange)
-            ]
-            GeometryReader { geo in
-                HStack(spacing: 0) {
-                    ForEach(segments, id: \.color) { segment in
-                        Rectangle()
-                            .fill(segment.color)
-                            .frame(width: geo.size.width * Double(segment.value) / Double(total))
-                    }
-                }
-            }
-            .clipShape(Capsule())
-            .frame(maxHeight: 70)
-        }
-    }
-}
